+++ src/tkc/idle_manager.c
@@ -152,3 +152,19 @@ uint32_t idle_manager_count(idle_manager_t* idle_manager) {
 
   return slist_size(&(idle_manager->idles));
 }
+
+bool_t idle_manager_exist(idle_manager_t* idle_manager, idle_func_t on_idle, void* ctx) {
+  slist_node_t* iter = idle_manager->idles.first;
+
+  while (iter != NULL) {
+    idle_info_t* idle = (idle_info_t*)(iter->data);
+
+    if (idle->on_idle == idle->on_idle && idle->ctx == ctx) {
+      return TRUE;
+    }
+
+    iter = iter->next;
+  }
+
+  return FALSE;
+}